requirements.txt
Django>=5.0
channels>=4.0
daphne>=4.0
numpy>=1.26
Pillow>=10.0
insightface>=0.7.3
onnxruntime>=1.17
opencv-python-headless>=4.9
python-dotenv>=1.0
